requirements.txt
amqp==2.6.1
appnope==0.1.2
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.4
backcall==0.2.0
billiard==3.6.4.0
boto3==1.17.38
botocore==1.20.38
celery==4.3.0
decorator==4.4.2
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-redis==4.12.1
django-rest-passwordreset==1.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.9.0
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.5.2
mccabe==0.6.1
model-bakery==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.11
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
redis==3.2.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
